+++ ruoyi-ui/package.json
@@ -1,7 +1,7 @@
 {
   "private": "true",
   "name": "@tencent/tdesign-vue-next-starter",
-  "version": "0.9.0",
+  "version": "0.10.0",
   "description": "Ruoyi-Tdesign多租户管理系统",
   "type": "module",
   "scripts": {
