+++ ruoyi-ui/src/pages/system/user/profile/BindingEmail.vue
