+++ ruoyi-ui/src/pages/system/user/profile/index.less
